Corrected inconsistent spacing, "fixed" race condition
Since message passing is asynchronous, there is a race condition
on disabling and then re-enabling the menu quickly that was causing
issues in my browser. This fix has consistent behavior, but the
first click on a matching element will not enable the menu before
it becomes visisble, and the first click on a not-matching element
will not hide the menu before it becomes visible.